requirements.txt
pillow
odfpy